+++ src/Support/Privacy.php
@@ -8,13 +8,18 @@ abstract class Privacy
 {
     public function fire(): bool
     {
-        return $this->validate(func_get_args()[0]);
+        $queryArgs = func_get_arg(0);
+        $queryContext = func_get_arg(1);
+
+        return $this->validate($queryArgs, $queryContext);
     }
 
     /**
-     * @param  array  $queryArgs  Arguments given with the query/mutation
+     * @param array $queryArgs    Arguments given with the query/mutation
+     * @param mixed $queryContext Context of the query/mutation
+     *
      * @return bool Return `true` to allow access to the field in question,
      *   `false otherwise
      */
-    abstract public function validate(array $queryArgs): bool;
+    abstract public function validate(array $queryArgs, $queryContext = null): bool;
 }