+++ tests/Database/SelectFields/ValidateFieldTests/PrivacyDenied.php
@@ -9,11 +9,9 @@ use Rebing\GraphQL\Support\Privacy;
 class PrivacyDenied extends Privacy
 {
     /**
-     * @param  array  $queryArgs Arguments given with the query/mutation
-     * @return bool Return `true` to allow access to the field in question,
-     *   `false otherwise
+     * @inheritDoc
      */
-    public function validate(array $queryArgs): bool
+    public function validate(array $queryArgs, $queryContext = null): bool
     {
         return false;
     }