+++ tests/Database/SelectFields/ValidateFieldTests/PrivacyQueryContext.php
@@ -7,17 +7,18 @@ namespace Rebing\GraphQL\Tests\Database\SelectFields\ValidateFieldTests;
 use PHPUnit\Framework\Assert;
 use Rebing\GraphQL\Support\Privacy;
 
-class PrivacyArgs extends Privacy
+class PrivacyQueryContext extends Privacy
 {
     /**
      * @inheritDoc
      */
     public function validate(array $queryArgs, $queryContext = null): bool
     {
-        $expectedQueryArgs = [
-            'arg_from_query' => true,
+        $expectedQueryContext = [
+            'arg_from_context_true' => true,
+            'arg_from_context_false' => false,
         ];
-        Assert::assertSame($expectedQueryArgs, $queryArgs);
+        Assert::assertSame($expectedQueryContext, $queryContext);
 
         return true;
     }